Reenable native modules.
This is the second attempt at enabling native modules with the airbnb babel
preset.  The first attempt used strings instead of require statements, and
broke plugin resolution when babel-preset-airbnb was a sub dependency.

Passing the option `modules: false` defeats the purpose of the `without-strict`
preset, as strict is an option being applied to a plugin that `modules: false`
removes.  This change will enable users to transform `import` and `import()`
statements with webpack2 instead of babel, and will enable tree shaking.