Switches to use babel-preset-env
Babel7 is set to deprecate the ES2015 preset and the new recommended
approach is to use babel-preset-env which also lets you fine tune the
transforms specific to the environments the code will be run on. This
commit moves our implementation to use babel-preset-env as well as lists
a few browsers we support.