Add looseClasses option to compile classes in loose mode
Compiling classes in normal mode is safer, but it comes with a small
bundle size and runtime overhead. For consumers that don't want to make
this tradeoff, I am adding an option to disable this. Poking around
other large codebases and I have been unable to find any other company
not using loose mode for classes.

  https://babeljs.io/docs/en/babel-plugin-transform-classes#loose

The two main issues outlined in the docs regarding loose mode are that
loose mode class methods are enumerable and there is an edge case where
extending a class that has a setter with the same name as an instance
method, the setter will be triggered.