Default jscript option to use explorer target versions
Instead of providing an explicit option for disabling this
transformation, we can make this automatic by using the target browser
versions that are affected by providing this transformation.

I didn't remove the option entirely yet because that would be a breaking
change, so I left it in but removed the documentation. We can remove it
in the next major version.